Fix content type matching, package references, and example issues

+++ examples/SampleApi/Program.cs
@@ -51,5 +51,5 @@ app.Run();
 
 record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
 {
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+    public int TemperatureF => 32 + (int)((double)TemperatureC / 0.5556);
 }
